chore: move documentation script files to /scripts/docs
