+++ scripts/docs/home.js
