+++ scripts/docs/tutorialsIndex.js
@@ -2,10 +2,12 @@
 
 const fs = require('fs');
 
-const tutorials = fs.readdirSync(__dirname).filter(file => file.endsWith('.md'));
+const dirName = "./docs/tutorials/"
+
+const tutorials = fs.readdirSync(dirName).filter(file => file.endsWith('.md'));
 
 module.exports = tutorials.reduce((map, filename) => {
-  const content = fs.readFileSync(`${__dirname}/${filename}`, 'utf8');
+  const content = fs.readFileSync(`${dirName}/${filename}`, 'utf8');
   map[`docs/tutorials/${filename}`] = {
     title: `Mongoose Tutorials: ${content.split('\n')[0].replace(/^#+/, '').trim()}`,
     acquit: true,
