+++ scripts/docs/typescriptIndex.js
@@ -2,10 +2,12 @@
 
 const fs = require('fs');
 
-const tutorials = fs.readdirSync(__dirname).filter(file => file.endsWith('.md'));
+const dirName = "./docs/typescript/";
+
+const tutorials = fs.readdirSync(dirName).filter(file => file.endsWith('.md'));
 
 module.exports = tutorials.reduce((map, filename) => {
-  const content = fs.readFileSync(`${__dirname}/${filename}`, 'utf8');
+  const content = fs.readFileSync(`${dirName}/${filename}`, 'utf8');
   map[`docs/typescript/${filename}`] = {
     title: `Mongoose: ${content.split('\n')[0].replace(/^#+/, '').trim()}`,
     markdown: true
