+++ scripts/website.js
@@ -106,8 +106,8 @@ try {
   fs.mkdirSync(path.join(cwd, './docs/api'));
 } catch (err) {} // eslint-disable-line no-empty
 
-require('../docs/splitApiDocs');
-const filemap = Object.assign({}, require('../docs/source'), require('../docs/tutorials'), require('../docs/typescript'));
+require('./docs/splitApiDocs');
+const filemap = Object.assign({}, require('./docs/docsIndex'), require('./docs/tutorialsIndex'), require('./docs/typescriptIndex'));
 const files = Object.keys(filemap);
 
 const wrapMarkdown = (md, baseLayout) => `
